Fetch repository information
This demos a tightly coupled integration with a web service (GitHub in this
case) with a very wild dependency graph between the fragment, the activity, the
adapter and the loader. I wonder if there's a pattern to untangle this mess? :)

This also adds RetroFit to the mix. Cause, why not?