Add tests to SillyUsernameFormatter
Non-determinism and side-effects make unit testing incredibly painful. If we
carefully inject all our side-effecting dependencies, we can easily mock them
out while testing. For this case, we inject a new Random instance through the
constructor and provide it via Dagger during runtime, but replace it with a fake
instance for testing where we can control the result.