Use Mockito to test the SillyUsernameFormatter
Mockito can save us from writing a ton of boilerplate code. In this case it
doesn't seem to matter a lot, but that is because we extend the Random class and
don't have to implement a complete interface. In addition to creating simple
no-op mocks, Mockito also enables us to record and verify the behavior of mocks
which we will see later.

To run the tests, start an emulator or connect a device and run

`./gradlew connectedAndroidTest`