Replace loaders with rxjava
Loaders are great compared to AsyncTasks because they're independent of the
activity lifecycle. So we don't have to worry about leaking stuff on
configuration changes and we also keep the data we've already fetched and don't
trigger unnecessary queries.

However, they're incredibly clunky. You need a lot of boilerplate to get them
set up and doing it right is quite tricky. RxJava lets us do this in a very
nice, concise and intuitive way and it's natively supported by RetroFit which
leads to some really well readable code.

RxJava is way more powerful than just this, but let's just start with a simple
example like this.